improve(stage-set): renamed primitives/Controls to primitives and mesh to actors.
