+++ src/fp-render/stage-set.ts
@@ -1,7 +1,7 @@
 import * as THREE from 'three';
 
-import * as MESH from './Mesh';
-import * as CONTROLS from './primitives/Controls';
+import * as MESH from './actors';
+import * as CONTROLS from './primitives';
 
 import * as E from 'fp-ts/lib/Either';
 import * as F from 'fp-ts/lib/function';
